+++ tee-worker/litentry/core/identity-verification/Cargo.toml
@@ -6,7 +6,6 @@ version = "0.1.0"
 
 [dependencies]
 # std dependencies
-base64 = { version = "0.22", default-features = false, features = ["alloc"] }
 hex = { version = "0.4.3", optional = true }
 rand = { version = "0.8", optional = true }
